list; use flexbox for all the header items
The commit efccc73d179 introduced the usage of flexbox to properly align
the filter form. It also introduced a regression with mobile device.

The issue was due to a mixed usage of flexbox and floating div.

This commit add flexbox layout for the whole list header:
* title
* add filter/create/refresh buttons
* filter form.

Some small changes are observed, the main one is the alignement of the
"Refresh" button on the right on mobile.